Switched out notification modules
For issue #2

I switched out notification modules to make this program support Linux and MacOS devices. Although I don't own devices of either OS, it works on Windows still.

I also changed the paths to search for the current directory instead of the absolute path, if ran via .exe it would result in a crash because it cannot find any of the files.